requirements.txt
streamlit
pandas
yfinance
plotly
pytesseract
pillow
scikit-learn
pandas_ta